A bunch of part changes to comply with the latest MTS code.
DO NOT TEST IN RELEASED VERSIONS
Reduced the RPM on the AEIO-540 to match the real one, and increased the pitch on the VTA-1's prop to make up for the speed increase. VT-60 prop got a significant pitch increase due to the lower prop feedback we have now, allowing it to reach upwards of 350 MPH and cruise at around 330 at a safe RPM when the jet thrust is included as well. PT6A-67B had a starter power change to keep the sounds working.